Fair Twin Inspection convenience function.

Fair Twin Inspection convenience function.

fairTwins(x, train.id = seq_len(nrow(x$train)), test.id = NULL, cols = NULL)

Arguments

x

Object of class fairadapt, a result of an adaptation procedure.

train.id

A vector of indices specifying which rows of the training data should be displayed.

test.id

A vector of indices specifying which rows of the test data should be displayed.

cols

A character vector, subset of names(train.data), which specifies which subset of columns is to be displayed in the result.

Value

A data.frame, containing the original and adapted values of the requested individuals. Adapted columns have _adapted appended to their original name.

Examples

n_samp <- 200
uni_dim <- c(       "gender", "edu", "test", "score")
uni_adj <- matrix(c(       0,     1,      1,       0,
                           0,     0,      1,       1,
                           0,     0,      0,       1,
                           0,     0,      0,       0),
                  ncol = length(uni_dim),
                  dimnames = rep(list(uni_dim), 2),
                  byrow = TRUE)

uni_ada <- fairadapt(score ~ .,
  train.data = head(uni_admission, n = n_samp),
  test.data = tail(uni_admission, n = n_samp),
  adj.mat = uni_adj,
  prot.attr = "gender"
)

fairTwins(uni_ada, train.id = 1:5)
#>   gender      score score_adapted        edu edu_adapted         test
#> 1      1  1.9501728     1.7131166  1.3499572  0.79837621  1.617739679
#> 2      0 -2.3502495    -2.3502495 -1.9779234 -1.97792341 -3.121796235
#> 3      1  0.6285619     0.0529272  0.6263626  0.07040801  0.530034686
#> 4      1  0.7064857     0.2603615  0.8142112  0.24114051  0.004573003
#> 5      1  0.3678313     0.0529272  1.8415242  0.84381181  1.193677123
#>   test_adapted
#> 1    1.0915941
#> 2   -3.1217962
#> 3    0.1064045
#> 4   -0.3906711
#> 5    0.8075426