Prediction function for new data from a saved fairadapt object.

Prediction function for new data from a saved fairadapt object.

# S3 method for fairadapt
predict(object, newdata, ...)

Arguments

object

Object of class fairadapt, a result of an adaptation procedure.

newdata

A data.frame containing the new data.

...

Additional arguments forwarded to computeQuants().

Value

A data.frame containing the adapted version of the new data.

Details

The newdata argument should be compatible with adapt.test argument that was used when constructing the fairadapt object. In particular, newdata should contain column names that appear in the formula argument that was used when calling fairadapt() (apart from the outcome variable on the LHS of the formula).

Examples

n_samp <- 200
uni_dim <- c(       "gender", "edu", "test", "score")
uni_adj <- matrix(c(       0,     1,      1,       0,
                           0,     0,      1,       1,
                           0,     0,      0,       1,
                           0,     0,      0,       0),
                  ncol = length(uni_dim),
                  dimnames = rep(list(uni_dim), 2),
                  byrow = TRUE)

uni_ada <- fairadapt(score ~ .,
  train.data = head(uni_admission, n = n_samp),
  adj.mat = uni_adj,
  prot.attr = "gender"
)

predict(object = uni_ada, newdata = tail(uni_admission, n = n_samp))
#>      gender          edu         test        score
#> 801       0 -1.762386175 -1.211890953 -2.248523253
#> 802       0 -1.455416223 -2.361166481 -2.009755697
#> 803       0  1.170794530  1.169657867  1.191612986
#> 804       0  0.440124823 -0.229145813 -0.697736661
#> 805       0 -1.014875674 -0.559947522  1.984450007
#> 806       0 -0.430447496 -1.209261161  0.789841297
#> 807       0 -0.332573698 -1.240123149 -1.474926656
#> 808       0 -0.544864823  0.812795420  0.497834190
#> 809       0 -0.821708580 -1.353303876  0.982337819
#> 810       0 -1.996778408 -2.023746740 -1.670251654
#> 811       0 -0.346160022 -0.686068562  1.872163467
#> 812       0  0.097568582 -0.046012506  1.109306522
#> 813       0 -1.059961139 -0.364127628 -0.347495108
#> 814       0 -1.393967432 -1.215501088  0.841504242
#> 815       0 -0.424532919 -1.168627410 -1.031754054
#> 816       0 -1.094412168 -2.603903781 -1.038394788
#> 817       0 -1.320001012 -2.665623320 -1.253841133
#> 818       0 -1.317427229 -0.838020739 -1.127569653
#> 819       0 -1.177133779 -2.072601121 -1.321760978
#> 820       0  0.008667341 -0.985364068  1.062412448
#> 821       0 -1.779031462 -2.603903781 -1.627486071
#> 822       0 -0.592176006  0.124843632  0.871076451
#> 823       0  0.008667341 -0.229145813  0.667045683
#> 824       0 -0.289977458 -1.168627410 -1.783792760
#> 825       0  0.620924161 -0.114971666  0.050424648
#> 826       0 -1.307013160 -1.483878994 -1.244894150
#> 827       0 -0.313852640 -0.781372572  0.836772353
#> 828       0  0.566041719  0.807542567 -0.318456042
#> 829       0  1.663760353  1.480402631  2.299315681
#> 830       0 -1.958414006 -3.121796235 -0.780448031
#> 831       0 -1.094412168 -2.361166481  0.143680389
#> 832       0 -0.562262231 -0.839388153 -0.197988253
#> 833       0  0.675487598  0.868880270 -0.151364462
#> 834       0 -0.869229220 -2.646687427 -1.604124109
#> 835       0 -0.256472779 -1.440056662 -1.924364325
#> 836       0 -0.015333931 -0.820665352  1.861951160
#> 837       0 -1.779031462 -1.942847492 -1.462275080
#> 838       0  0.138359545 -1.790409294 -0.520312886
#> 839       0  0.229347734  0.106404515  0.318168927
#> 840       0  0.477895250  0.229617350 -0.422195985
#> 841       0 -1.268711009 -3.497560395 -0.094699039
#> 842       0  0.844084462 -0.305803234 -0.789176231
#> 843       0 -0.346160022 -0.781372572 -1.156295463
#> 844       0  0.799480614  0.539023997  0.988820130
#> 845       0 -0.094751785 -1.184286201 -0.383527997
#> 846       0  0.877554170 -0.005874030  1.297151367
#> 847       0 -1.043859083 -1.920065994 -0.509722245
#> 848       0  1.184813575  0.223671853  0.948738536
#> 849       0  0.509482542  0.031115903 -0.960078119
#> 850       0 -0.295124669 -0.004056565  0.069640746
#> 851       0 -0.280587143 -1.968535319 -0.642857333
#> 852       0  0.241140510  0.855783424  1.190722501
#> 853       0  0.799480614  0.199833484  1.672520558
#> 854       0 -1.043859083 -2.867721330 -1.376894259
#> 855       0  1.511649979  0.665861758  1.724080192
#> 856       0 -0.716919789  0.041877886  0.169921945
#> 857       0 -0.439514510 -2.782963224 -1.776423466
#> 858       0 -1.283178362 -1.390920653 -1.176246440
#> 859       0  0.279666993 -0.187664519  1.105173350
#> 860       0 -0.346160022 -1.559090289 -0.352586641
#> 861       0 -1.393405437 -2.227771994 -1.211904179
#> 862       0  0.650752035  0.870076964  0.172484884
#> 863       0 -0.257847380  0.184278366 -0.955779822
#> 864       0 -0.037313912 -0.556922603 -1.240644748
#> 865       0 -1.317427229 -0.629011687 -0.648786856
#> 866       0  0.836348371  1.801792786  0.816701249
#> 867       0 -0.659880020 -1.844452022 -0.468480289
#> 868       0  0.236144107  0.386751900  0.929639751
#> 869       0  1.656675158  1.174964842  0.453241313
#> 870       0 -2.022081286 -1.905372715 -1.723788786
#> 871       0  0.699699390  0.029302486 -0.035326898
#> 872       0 -0.284292292 -0.497271008  0.608830823
#> 873       0 -1.167368621 -1.865114023 -2.006507850
#> 874       0  0.200417113  0.344140793  0.113582414
#> 875       0  0.316291286 -0.624238187  1.462430907
#> 876       0  0.623657245 -0.972542179  1.586366410
#> 877       0  0.253398410 -0.629011687 -0.142862516
#> 878       0 -0.346160022 -0.550042638  1.685246091
#> 879       0 -1.177133779 -2.036705661 -0.172807492
#> 880       0 -2.179599761 -2.983597500 -0.935532589
#> 881       0 -1.725707315 -2.820512900 -3.559906963
#> 882       0 -1.557369432 -2.772926068 -2.731656754
#> 883       0 -0.431383366 -1.168627410  0.094147384
#> 884       0  0.896131428  0.645716214  2.973083804
#> 885       0  0.097557484 -0.393229676  0.856685667
#> 886       0 -1.743314709 -3.082900705 -2.279216961
#> 887       0 -2.344488163 -1.435563283 -2.255204660
#> 888       0 -1.958414006 -1.969555323 -1.624173997
#> 889       0  0.140822667 -1.434405915  0.095045586
#> 890       0 -3.069242231 -3.712797683 -4.430320178
#> 891       0 -0.562262231 -1.559090289 -1.282782046
#> 892       0 -0.730197066 -1.924209626 -0.867176978
#> 893       0 -0.103882909 -0.546331814 -1.116750069
#> 894       0  0.498140800 -0.395424735 -0.443540996
#> 895       0  0.513260622  0.674104083  0.954496259
#> 896       0 -0.054786180 -1.120835000  0.533034913
#> 897       0  0.824376597  0.838675786  2.434167705
#> 898       0 -0.562262231 -0.170059038  1.738022171
#> 899       0 -0.440428486 -1.486504533 -2.522900649
#> 900       0  0.697135387 -0.825026405 -1.421399698
#> 901       0  0.793304107  2.024935105  0.876393816
#> 902       0 -2.004058295 -2.665623320 -3.336136856
#> 903       0 -1.005181934 -1.465181188 -1.620118314
#> 904       0 -1.763352581 -2.665623320 -1.298079833
#> 905       0 -1.314466013 -2.279801899 -0.194242511
#> 906       0 -0.440428486 -0.985364068 -0.539844056
#> 907       0  0.279666993  0.190043735  1.540468960
#> 908       0 -0.288401907 -2.072601121  0.377569531
#> 909       0 -1.317427229 -2.023746740 -0.518650912
#> 910       0  0.140822667 -1.168627410 -1.067862289
#> 911       0 -1.316093279 -1.754439190 -3.002032488
#> 912       0 -0.423314434 -1.375426561 -1.039456538
#> 913       0  0.489578511 -1.241887015  0.657435765
#> 914       0 -1.177133779 -1.533112500 -0.951787515
#> 915       0 -0.118696696 -1.566638938 -1.001299575
#> 916       0 -0.362287300 -1.498664881 -1.546853316
#> 917       0 -0.251867933  0.107841476  0.952588498
#> 918       0 -1.331568894 -1.995935302 -1.853927038
#> 919       0 -0.717063092 -1.275608741 -1.474532010
#> 920       0 -1.779031462 -2.276587625 -0.917686286
#> 921       0 -0.820177528 -1.582146965 -2.008219189
#> 922       0 -1.854995822 -2.496803403 -1.438337544
#> 923       0 -0.274171366 -0.933150449 -0.692028282
#> 924       0 -0.638266370 -1.957392080 -0.837267282
#> 925       0 -1.736567678 -2.390473070 -1.981547684
#> 926       0  0.279666993 -0.070526087  1.517489869
#> 927       0 -0.560277832 -0.356605410  0.631959853
#> 928       0 -0.274171366 -0.064637904  0.945444557
#> 929       0 -0.507195135 -1.840148830 -1.432336509
#> 930       0 -0.562262231 -1.168627410  1.250912889
#> 931       0 -2.927492507 -2.892119311 -2.289800236
#> 932       0  0.427196470  0.433596689  0.899829891
#> 933       0 -1.610681784 -1.443623594 -1.207835484
#> 934       0 -0.884406639 -0.440364450  0.465921575
#> 935       0 -0.638071046 -1.156573500 -0.280759713
#> 936       0  0.868146498  1.571927196  1.664871387
#> 937       0 -1.524549736 -0.370525777 -0.948986281
#> 938       0  0.896131428 -0.501127666  1.682221972
#> 939       0 -2.455354703 -3.121796235 -2.162776061
#> 940       0 -0.288401907 -0.686068562  0.587418523
#> 941       0 -1.331568894 -0.946489317  0.356856675
#> 942       0  0.065275309 -0.658202059 -0.317153475
#> 943       0 -1.455416223 -2.072601121 -1.481437987
#> 944       0 -0.560277832 -0.693222303 -1.990669444
#> 945       0  0.957810325  0.915731389  2.498366165
#> 946       0  0.280579357 -0.264326178  0.324320914
#> 947       0  1.806698793  1.084950383  0.895222646
#> 948       0  1.348934868  1.173632593  1.802367455
#> 949       0  0.595814141  1.926888054  0.364802173
#> 950       0  0.014562394 -0.342731345  0.381268872
#> 951       0 -0.672490266 -1.095405615  0.569697590
#> 952       0 -0.100453020 -0.818827187 -0.549087261
#> 953       0 -0.311265518 -0.682830213  0.166889733
#> 954       0 -1.317427229 -1.480622677 -0.911256296
#> 955       0  0.319238271 -0.501127666  0.213121286
#> 956       0  0.665089087  0.568883700  0.150221197
#> 957       0 -1.598016256 -2.266070448 -2.750619416
#> 958       0 -1.390880951 -1.755829534 -1.691135270
#> 959       0  0.243085799 -0.227891531 -1.326310399
#> 960       0 -0.600479913  0.288362252 -0.759343490
#> 961       0 -0.380612020 -1.420420046 -1.300735252
#> 962       0  0.474592506  0.036977119  0.829957649
#> 963       0  1.012320032  0.544826681  0.149567854
#> 964       0 -1.445846363 -1.515208888 -0.470509409
#> 965       0 -0.598164281 -0.962079122 -1.205855074
#> 966       0 -1.331568894 -1.969555323  0.617005511
#> 967       0 -0.834599062 -1.340114903 -1.219394858
#> 968       0 -1.317427229 -0.946489317  0.091312402
#> 969       0 -0.460064592  0.033031686  0.118477910
#> 970       0 -0.463552976 -0.549713195 -0.321729818
#> 971       0 -0.054786180 -1.995935302  1.218577031
#> 972       0 -1.006788858 -1.272298519  0.333778613
#> 973       0 -0.207325249 -0.790899405 -0.164120438
#> 974       0 -0.198671678 -1.424801374 -0.515737721
#> 975       0 -1.320001012 -1.373614666 -0.585724370
#> 976       0 -1.393967432 -2.603903781 -0.554345955
#> 977       0 -1.094412168 -2.867721330 -1.142323022
#> 978       0 -0.302652231 -0.403718017 -0.025212237
#> 979       0 -1.391352979 -1.955211876 -1.182021588
#> 980       0 -1.014875674 -2.036466214 -0.504013446
#> 981       0  0.953697789  0.506355585  0.524688786
#> 982       0 -0.930674470 -2.383798643 -3.030736973
#> 983       0  1.149041867 -0.124322538 -0.214029243
#> 984       0 -0.638071046 -0.686068562  0.531710091
#> 985       0 -2.004058295 -2.665623320 -2.195556550
#> 986       0 -0.882122467 -1.434405915 -0.231555284
#> 987       0  0.097568582  0.107841476  2.017461669
#> 988       0 -0.562262231 -1.168627410  1.075378490
#> 989       0 -0.638266370 -0.795081657  1.104165903
#> 990       0 -0.274171366 -0.755109628  1.739168447
#> 991       0  0.460179935 -0.490180198  0.100981820
#> 992       0  0.213541419 -1.815610955 -2.239312858
#> 993       0 -1.574354070 -1.072093592 -1.288697389
#> 994       0 -0.562262231 -1.124719294  2.041889298
#> 995       0 -1.238780503 -2.409644152 -3.071024177
#> 996       0  0.111594251  0.682887893 -2.155281612
#> 997       0 -0.050573998 -0.550042638 -0.789234909
#> 998       0 -1.304409643 -0.985364068  0.049963241
#> 999       0 -1.455416223 -1.955211876 -0.360933334
#> 1000      0 -0.867018060 -1.689133889 -0.004731789