fairadapt

Implementation of fair data adaptation with quantile preservation (Plecko & Meinshausen 2019). Uses only plain R.

fairadapt(
  formula,
  train.data,
  test.data = NULL,
  adj.mat = NULL,
  cfd.mat = NULL,
  top.ord = NULL,
  protect.A,
  res.vars = NULL,
  quant.method = rangerQuants,
  visualize.graph = FALSE,
  ...
)

Arguments

formula
Object of class formula describing the response and the covariates.

train.data, test.data
Training data & testing data, both of class data.frame. Test data is by default NULL.

adj.mat
Matrix of class matrix encoding the relationships in the causal graph. M[i,j] == 1L implies the existence of an edge from node i to node j. Must include all the variables appearing in the formula object.

cfd.mat
Symmetric matrix of class matrix encoding the bidirected edges in the causal graph. M[i,j] == M[j, i] == 1L implies the existence of a bidirected edge between nodes i and j. Must include all the variables appearing in the formula object.

top.ord
A vector of class character describing the topological ordering of the causal graph. Default value is NULL, but this argument must be supplied if adj.mat is not specified. Also must include all the variables appearing in the formula object.

protect.A
A value of class character describing the binary protected attribute. Must be one of the entries of colnames(adj.mat).

res.vars
A vector of class character listing all the resolving variables, which should not be changed by the adaption procedure. Default value is NULL, corresponding to no resolving variables. Resolving variables should be a subset of the descendants of the protected attribute.

quant.method
A vector of class character choosing the method used for quantile regression. Possible values are "forest", "forest2", "linear" and "nn".

visualize.graph
A logical indicating whether the causal graph should be plotted upon calling the fairadapt() function. Default value is FALSE.

...
Forwarded to the function passed as `quant.method`.

Value

An object of class fairadapt, containing the original and adapted training and testing data, together with the causal graph and some additional meta-information.

Details

The procedure takes the training and testing data as an input, together with the causal graph given by an adjacency matrix and the list of resolving variables, which should be kept fixed during the adaptation procedure. The procedure then calculates a fair representation of the data, after which any classification method can be used. There are, however, several valid training options yielding fair predictions, and the best of them can be chosen with cross-validation. For more details we refer the user to the original paper. Most of the running time is due to the quantile regression step using the ranger package.

References

Plecko, D. & Meinshausen, N. (2019). Fair Data Adaptation with Quantile Preservation

Author

Drago Plecko

Examples

uni.adj.mat <- array(0, dim = c(4, 4))
colnames(uni.adj.mat) <- rownames(uni.adj.mat) <-
  c("gender", "edu", "test", "score")

uni.adj.mat["gender", c("edu", "test")] <-
  uni.adj.mat["edu", c("test", "score")] <-
  uni.adj.mat["test", "score"] <- 1L

FA <- fairadapt(score ~ .,
  train.data = uni_admission[1:100, ],
  test.data = uni_admission[101:150, ],
  adj.mat = uni.adj.mat, protect.A = "gender")

FA

#> Fairadapt result
#> 
#> Call:
#>  score ~ . 
#> 
#> Protected attribute:                  gender 
#> Protected attribute levels:           0, 1 
#> Number of training samples:           100 
#> Number of test samples:               50 
#> Number of independent variables:      3 
#> Total variation (before adaptation):  -0.573929 
#> Total variation (after adaptation):   0.06771308