Fair Twin Inspection convenience function.

Fair Twin Inspection convenience function.

fairTwins(x, train.id = 1L, test.id = NULL, cols = NULL)

Arguments

x
Object of class fairadapt, a result of an adaptation procedure.

train.id
A vector of indices specifying which rows of the training data should be displayed.

test.id
A vector of indices specifying which rows of the test data should be displayed.

cols
A character vector, subset of names(train.data), which specifies which subset of columns is to be displayed in the result.

Value

A data.frame, containing the original and adapted values of the requested individuals. Adapted columns have _adapted appended to their original name.

Examples

uni.adj.mat <- array(0, dim = c(4, 4))
colnames(uni.adj.mat) <- rownames(uni.adj.mat) <-
  c("gender", "edu", "test", "score")

uni.adj.mat["gender", c("edu", "test")] <-
  uni.adj.mat["edu", c("test", "score")] <-
  uni.adj.mat["test", "score"] <- 1L

FA <- fairadapt(score ~ .,
  train.data = uni_admission[1:100, ],
  test.data = uni_admission[101:150, ],
  adj.mat = uni.adj.mat, prot.attr = "gender")

fairTwins(FA, train.id = 1:5)

#>   gender      score score_adapted        edu edu_adapted         test
#> 1      1  1.9501728    1.43925785  1.3499572   0.1488319  1.617739679
#> 2      0 -2.3502495   -2.35024955 -1.9779234  -1.9779234 -3.121796235
#> 3      1  0.6285619    0.01747699  0.6263626  -0.4181586  0.530034686
#> 4      1  0.7064857    0.18543763  0.8142112  -0.2899775  0.004573003
#> 5      1  0.3678313    0.01747699  1.8415242   0.7144597  1.193677123
#>   test_adapted
#> 1   0.56088179
#> 2  -3.12179624
#> 3  -0.18189652
#> 4  -1.12083500
#> 5   0.03774411