Prediction function for new data from a saved fairadaptBoot object.

Prediction function for new data from a saved fairadaptBoot object.

# S3 method for fairadaptBoot
predict(object, newdata, ...)

Arguments

object

Object of class fairadapt, a result of an adaptation procedure.

newdata

A data.frame containing the new data.

...

Additional arguments forwarded to computeQuants().

Value

A data.frame containing the adapted version of the new data.

Details

The newdata argument should be compatible with adapt.test argument that was used when constructing the fairadaptBoot object. In particular, newdata should contain column names that appear in the formula argument that was used when calling fairadaptBoot() (apart from the outcome variable on the LHS of the formula).

Examples

n_samp <- 200
uni_dim <- c(       "gender", "edu", "test", "score")
uni_adj <- matrix(c(       0,     1,      1,       0,
                           0,     0,      1,       1,
                           0,     0,      0,       1,
                           0,     0,      0,       0),
                  ncol = length(uni_dim),
                  dimnames = rep(list(uni_dim), 2),
                  byrow = TRUE)

uni_ada_boot <- fairadaptBoot(score ~ .,
  train.data = head(uni_admission, n = n_samp),
  adj.mat = uni_adj,
  prot.attr = "gender",
  n.boot = 5,
  keep.object = TRUE
)

predict(object = uni_ada_boot, newdata = tail(uni_admission, n = n_samp))
#> [[1]]
#>      gender         edu          test        score
#> 801       0 -1.76238617 -1.2118909528 -2.248523253
#> 802       0 -1.71380342 -2.0726011207 -2.009755697
#> 803       0  1.17079453  1.1696578669  1.191612986
#> 804       0  0.28765109 -0.1876645189 -0.697736661
#> 805       0 -0.86164724 -0.5011276663  1.984450007
#> 806       0 -0.43044750 -1.2092611606  0.789841297
#> 807       0 -0.33257370 -1.3736146655 -1.474926656
#> 808       0 -0.54486482  0.8127954201  0.497834190
#> 809       0 -0.82170858 -1.3533038756  0.982337819
#> 810       0 -1.99677841 -1.9959353015 -1.670251654
#> 811       0 -0.34616002 -0.7307073976  1.872163467
#> 812       0 -0.05057400 -0.0646379037  1.109306522
#> 813       0 -1.05996114 -0.3641276278 -0.347495108
#> 814       0 -1.31742723 -1.2155010885  0.841504242
#> 815       0 -0.41815857 -1.3147835907 -1.031754054
#> 816       0 -0.97207326 -2.0383347537 -1.038394788
#> 817       0 -1.17098939 -2.6039037812 -1.253841133
#> 818       0 -1.10801481 -0.6714620022 -1.127569653
#> 819       0 -1.00303738 -1.6322041223 -1.321760978
#> 820       0 -0.05057400 -1.1340844308  1.062412448
#> 821       0 -1.97792341 -2.9835975002 -1.627486071
#> 822       0 -0.59217601  0.1248436319  0.871076451
#> 823       0 -0.05057400 -0.3169367808  0.667045683
#> 824       0 -0.28997746 -1.3147835907 -1.783792760
#> 825       0  0.62092416 -0.1149716659  0.050424648
#> 826       0 -1.10801481 -1.2155010885 -1.244894150
#> 827       0 -0.29205193 -0.8102336998  0.836772353
#> 828       0  0.56604172  0.8075425675 -0.318456042
#> 829       0  1.66376035  1.4804026310  2.299315681
#> 830       0 -1.97879074 -3.1180112726 -0.780448031
#> 831       0 -0.97207326 -2.0003295088  0.143680389
#> 832       0 -0.52640406 -0.8250264053 -0.197988253
#> 833       0  0.67548760  0.8688802700 -0.151364462
#> 834       0 -0.86922922 -2.6466874272 -1.604124109
#> 835       0 -0.25647278 -1.4400566615 -1.924364325
#> 836       0 -0.01533393 -0.8206653519  1.861951160
#> 837       0 -1.97792341 -2.0237467402 -1.462275080
#> 838       0  0.13835955 -1.7904092940 -0.520312886
#> 839       0  0.22934773  0.1064045150  0.318168927
#> 840       0  0.47789525  0.2296173496 -0.422195985
#> 841       0 -1.26871101 -3.4975603952 -0.094699039
#> 842       0  0.84408446 -0.3058032343 -0.789176231
#> 843       0 -0.34616002 -0.7950816572 -1.156295463
#> 844       0  0.79948061  0.5390239967  0.988820130
#> 845       0 -0.09475179 -1.1842862011 -0.383527997
#> 846       0  0.85743315 -0.0344526249  1.297151367
#> 847       0 -0.88212247 -1.1373335597 -0.509722245
#> 848       0  1.18481357  0.2236718526  0.948738536
#> 849       0  0.50948254  0.0311159033 -0.960078119
#> 850       0 -0.29512467 -0.0040565645  0.069640746
#> 851       0 -0.28058714 -1.9685353189 -0.642857333
#> 852       0  0.24114051  0.8557834238  1.190722501
#> 853       0  0.79948061  0.2432488918  1.672520558
#> 854       0 -0.88212247 -2.2831847757 -1.376894259
#> 855       0  1.35153831  1.0381595823  1.724080192
#> 856       0 -0.71691979  0.0418778859  0.169921945
#> 857       0 -0.43951451 -2.7829632242 -1.776423466
#> 858       0 -1.28317836 -1.3909206525 -1.176246440
#> 859       0  0.14082267 -0.1083605515  1.105173350
#> 860       0 -0.34616002 -1.9072514019 -0.352586641
#> 861       0 -1.39340544 -2.2277719943 -1.211904179
#> 862       0  0.65075203  0.8700769639  0.172484884
#> 863       0 -0.25784738  0.1842783660 -0.955779822
#> 864       0 -0.03731391 -0.5569226025 -1.240644748
#> 865       0 -1.10801481 -0.5599475220 -0.648786856
#> 866       0  0.83634837  1.8017927861  0.816701249
#> 867       0 -0.65988002 -1.8444520222 -0.468480289
#> 868       0  0.23614411  0.3867519004  0.929639751
#> 869       0  1.65667516  1.1749648419  0.453241313
#> 870       0 -2.02208129 -1.9053727150 -1.723788786
#> 871       0  0.71445973  0.0293024863 -0.035326898
#> 872       0 -0.28997746 -0.3566054098  0.608830823
#> 873       0 -1.16736862 -1.8651140231 -2.006507850
#> 874       0  0.20041711  0.3441407933  0.113582414
#> 875       0  0.14292108 -0.6860685622  1.462430907
#> 876       0  0.62365725 -0.9725421787  1.586366410
#> 877       0  0.11327945 -0.7569210412 -0.142862516
#> 878       0 -0.34616002 -0.5500426384  1.685246091
#> 879       0 -1.00303738 -1.4838789943 -0.172807492
#> 880       0 -2.00405830 -2.8677213300 -0.935532589
#> 881       0 -1.72570731 -2.8205128999 -3.559906963
#> 882       0 -1.55736943 -2.7729260675 -2.731656754
#> 883       0 -0.44042849 -1.2433001891  0.094147384
#> 884       0  0.89613143  0.9577863382  2.973083804
#> 885       0  0.09755748 -0.3932296764  0.856685667
#> 886       0 -1.74331471 -3.0829007048 -2.279216961
#> 887       0 -2.34448816 -1.4355632830 -2.255204660
#> 888       0 -1.97879074 -1.9695553229 -1.624173997
#> 889       0 -0.04401245 -1.5152088882  0.095045586
#> 890       0 -3.06924223 -3.7127976828 -4.430320178
#> 891       0 -0.52640406 -1.5590902885 -1.282782046
#> 892       0 -0.73019707 -1.9242096264 -0.867176978
#> 893       0 -0.10388291 -0.5463318140 -1.116750069
#> 894       0  0.49814080 -0.3954247351 -0.443540996
#> 895       0  0.51326062  0.6741040827  0.954496259
#> 896       0 -0.23377408 -1.3736146655  0.533034913
#> 897       0  0.82437660  0.8386757856  2.434167705
#> 898       0 -0.52640406  0.0003170373  1.738022171
#> 899       0 -0.45280658 -1.5152088882 -2.522900649
#> 900       0  0.69713539 -0.8393881530 -1.421399698
#> 901       0  0.79330411  2.0249351046  0.876393816
#> 902       0 -1.99701136 -2.5042331755 -3.336136856
#> 903       0 -1.00518193 -1.4651811884 -1.620118314
#> 904       0 -1.76991987 -2.6039037812 -1.298079833
#> 905       0 -1.31446601 -2.2798018987 -0.194242511
#> 906       0 -0.45280658 -0.9880673622 -0.539844056
#> 907       0  0.14082267  0.1254069315  1.540468960
#> 908       0 -0.28997746 -2.0661573038  0.377569531
#> 909       0 -1.10801481 -1.9959353015 -0.518650912
#> 910       0 -0.04401245 -1.4838789943 -1.067862289
#> 911       0 -1.31609328 -1.7544391897 -3.002032488
#> 912       0 -0.42331443 -1.3754265610 -1.039456538
#> 913       0  0.48957851 -1.2418870154  0.657435765
#> 914       0 -1.00303738 -1.0740894054 -0.951787515
#> 915       0 -0.11869670 -1.5666389377 -1.001299575
#> 916       0 -0.36228730 -1.4986648806 -1.546853316
#> 917       0 -0.28840191  0.2432488918  0.952588498
#> 918       0 -1.22256601 -1.9959353015 -1.853927038
#> 919       0 -0.71706309 -1.2756087410 -1.474532010
#> 920       0 -1.97792341 -2.6039037812 -0.917686286
#> 921       0 -0.82017753 -1.5821469651 -2.008219189
#> 922       0 -1.85499582 -2.4968034032 -1.438337544
#> 923       0 -0.28840191 -0.9748456421 -0.692028282
#> 924       0 -0.63795435 -1.9072514019 -0.837267282
#> 925       0 -1.73656768 -2.3904730700 -1.981547684
#> 926       0  0.14082267 -0.0058740295  1.517489869
#> 927       0 -0.46006459 -0.2444583886  0.631959853
#> 928       0 -0.28840191  0.0008095085  0.945444557
#> 929       0 -0.50719514 -1.8401488303 -1.432336509
#> 930       0 -0.52640406 -1.3313739538  1.250912889
#> 931       0 -2.92749251 -2.8921193105 -2.289800236
#> 932       0  0.42719647  0.4335966889  0.899829891
#> 933       0 -1.61068178 -1.4436235941 -1.207835484
#> 934       0 -0.88440664 -0.4403644500  0.465921575
#> 935       0 -0.62472615 -1.1036314851 -0.280759713
#> 936       0  0.86814650  1.5719271962  1.664871387
#> 937       0 -1.52454974 -0.3705257766 -0.948986281
#> 938       0  0.89613143 -0.3566054098  1.682221972
#> 939       0 -2.45535470 -3.1180112726 -2.162776061
#> 940       0 -0.28997746 -0.6860685622  0.587418523
#> 941       0 -1.22256601 -0.8393881530  0.356856675
#> 942       0  0.06527531 -0.6582020594 -0.317153475
#> 943       0 -1.71380342 -2.0657019592 -1.481437987
#> 944       0 -0.46006459 -0.5822566584 -1.990669444
#> 945       0  0.95781033  0.9157313885  2.498366165
#> 946       0  0.28057936 -0.2643261784  0.324320914
#> 947       0  1.80669879  1.0849503835  0.895222646
#> 948       0  1.14755867  1.2953057903  1.802367455
#> 949       0  0.59581414  1.9268880540  0.364802173
#> 950       0  0.01456239 -0.3427313447  0.381268872
#> 951       0 -0.67249027 -1.0954056150  0.569697590
#> 952       0 -0.10045302 -0.8188271874 -0.549087261
#> 953       0 -0.31126552 -0.6828302131  0.166889733
#> 954       0 -1.10801481 -1.2663337296 -0.911256296
#> 955       0  0.14883187 -0.5686619514  0.213121286
#> 956       0  0.66508909  0.5688837002  0.150221197
#> 957       0 -1.59801626 -2.2660704480 -2.750619416
#> 958       0 -1.39088095 -1.7558295342 -1.691135270
#> 959       0  0.24308580 -0.2278915310 -1.326310399
#> 960       0 -0.60047991  0.2883622521 -0.759343490
#> 961       0 -0.38061202 -1.4204200461 -1.300735252
#> 962       0  0.47459251  0.0369771194  0.829957649
#> 963       0  1.01232003  0.5448266812  0.149567854
#> 964       0 -1.43119017 -1.4983050310 -0.470509409
#> 965       0 -0.59816428 -0.9620791225 -1.205855074
#> 966       0 -1.22256601 -1.9552118762  0.617005511
#> 967       0 -0.83459906 -1.3401149034 -1.219394858
#> 968       0 -1.10801481 -0.8250264053  0.091312402
#> 969       0 -0.46006459  0.1099141997  0.118477910
#> 970       0 -0.46355298 -0.5497131954 -0.321729818
#> 971       0 -0.23377408 -2.0383347537  1.218577031
#> 972       0 -1.00678886 -1.2722985194  0.333778613
#> 973       0 -0.20732525 -0.7908994052 -0.164120438
#> 974       0 -0.19867168 -1.4248013741 -0.515737721
#> 975       0 -1.17098939 -1.2155010885 -0.585724370
#> 976       0 -1.31742723 -2.5042331755 -0.554345955
#> 977       0 -0.97207326 -2.0663504803 -1.142323022
#> 978       0 -0.30265223 -0.4037180167 -0.025212237
#> 979       0 -1.30667470 -1.9552118762 -1.182021588
#> 980       0 -0.86164724 -2.0361968366 -0.504013446
#> 981       0  0.95369779  0.5063555847  0.524688786
#> 982       0 -0.93067447 -2.3837986431 -3.030736973
#> 983       0  1.14904187 -0.1243225376 -0.214029243
#> 984       0 -0.62472615 -0.4896495155  0.531710091
#> 985       0 -1.99701136 -2.5042331755 -2.195556550
#> 986       0 -0.63897033 -1.3913657105 -0.231555284
#> 987       0 -0.05057400  0.1006712105  2.017461669
#> 988       0 -0.54015681 -1.3147835907  1.075378490
#> 989       0 -0.63795435 -0.7813725716  1.104165903
#> 990       0 -0.28840191 -0.7884464598  1.739168447
#> 991       0  0.46017993 -0.4901801981  0.100981820
#> 992       0  0.21354142 -1.8156109549 -2.239312858
#> 993       0 -1.57435407 -1.0720935917 -1.288697389
#> 994       0 -0.52640406 -1.1686274101  2.041889298
#> 995       0 -1.23878050 -2.4096441517 -3.071024177
#> 996       0  0.11159425  0.6828878935 -2.155281612
#> 997       0 -0.17080730 -0.5822566584 -0.789234909
#> 998       0 -1.04556375 -0.8951183210  0.049963241
#> 999       0 -1.71380342 -1.9552118762 -0.360933334
#> 1000      0 -0.86701806 -1.6891338893 -0.004731789
#> 
#> [[2]]
#>      gender          edu         test        score
#> 801       0 -1.762386175 -1.211890953 -2.248523253
#> 802       0 -1.445846363 -2.072601121 -2.009755697
#> 803       0  1.170794530  1.169657867  1.191612986
#> 804       0  0.440124823 -0.244458389 -0.697736661
#> 805       0 -1.001776420 -0.390671052  1.984450007
#> 806       0 -0.430447496 -1.209261161  0.789841297
#> 807       0 -0.332573698 -1.180720819 -1.474926656
#> 808       0 -0.544864823  0.812795420  0.497834190
#> 809       0 -0.821708580 -1.353303876  0.982337819
#> 810       0 -1.996778408 -2.007616106 -1.670251654
#> 811       0 -0.346160022 -0.686068562  1.872163467
#> 812       0  0.113279446 -0.046012506  1.109306522
#> 813       0 -1.059961139 -0.364127628 -0.347495108
#> 814       0 -1.391352979 -1.215501088  0.841504242
#> 815       0 -0.424559369 -1.181002061 -1.031754054
#> 816       0 -1.090163575 -2.361166481 -1.038394788
#> 817       0 -1.317427229 -2.504233176 -1.253841133
#> 818       0 -1.317427229 -0.686068562 -1.127569653
#> 819       0 -1.177133779 -2.065701959 -1.321760978
#> 820       0  0.008667341 -0.985364068  1.062412448
#> 821       0 -1.770521236 -2.276587625 -1.627486071
#> 822       0 -0.592176006  0.124843632  0.871076451
#> 823       0  0.008667341 -0.212029256  0.667045683
#> 824       0 -0.289977458 -1.159948595 -1.783792760
#> 825       0  0.620924161 -0.114971666  0.050424648
#> 826       0 -1.307013160 -1.434405915 -1.244894150
#> 827       0 -0.313933684 -0.755109628  0.836772353
#> 828       0  0.566041719  0.807542567 -0.318456042
#> 829       0  1.663760353  1.480402631  2.299315681
#> 830       0 -1.789435649 -2.690683473 -0.780448031
#> 831       0 -1.090163575 -2.072601121  0.143680389
#> 832       0 -0.560277832 -0.839388153 -0.197988253
#> 833       0  0.675487598  0.868880270 -0.151364462
#> 834       0 -0.869229220 -2.646687427 -1.604124109
#> 835       0 -0.256472779 -1.440056662 -1.924364325
#> 836       0 -0.015333931 -0.820665352  1.861951160
#> 837       0 -1.770521236 -1.907251402 -1.462275080
#> 838       0  0.138359545 -1.790409294 -0.520312886
#> 839       0  0.229347734  0.106404515  0.318168927
#> 840       0  0.477895250  0.229617350 -0.422195985
#> 841       0 -1.268711009 -3.497560395 -0.094699039
#> 842       0  0.844084462 -0.305803234 -0.789176231
#> 843       0 -0.346160022 -0.756921041 -1.156295463
#> 844       0  0.799480614  0.389377080  0.988820130
#> 845       0 -0.094751785 -1.184286201 -0.383527997
#> 846       0  0.879040351 -0.046012506  1.297151367
#> 847       0 -1.043859083 -1.504998958 -0.509722245
#> 848       0  1.184813575  0.223671853  0.948738536
#> 849       0  0.509482542  0.031115903 -0.960078119
#> 850       0 -0.295124669 -0.004056565  0.069640746
#> 851       0 -0.280587143 -1.968535319 -0.642857333
#> 852       0  0.241140510  0.855783424  1.190722501
#> 853       0  0.799480614  0.198769381  1.672520558
#> 854       0 -1.043859083 -2.608594466 -1.376894259
#> 855       0  1.511649979  0.696805763  1.724080192
#> 856       0 -0.716919789  0.041877886  0.169921945
#> 857       0 -0.439514510 -2.782963224 -1.776423466
#> 858       0 -1.283178362 -1.390920653 -1.176246440
#> 859       0  0.279666993 -0.187664519  1.105173350
#> 860       0 -0.346160022 -1.912519623 -0.352586641
#> 861       0 -1.393405437 -2.227771994 -1.211904179
#> 862       0  0.650752035  0.870076964  0.172484884
#> 863       0 -0.257847380  0.184278366 -0.955779822
#> 864       0 -0.037313912 -0.556922603 -1.240644748
#> 865       0 -1.317427229 -0.559947522 -0.648786856
#> 866       0  0.836348371  1.801792786  0.816701249
#> 867       0 -0.659880020 -1.844452022 -0.468480289
#> 868       0  0.236144107  0.386751900  0.929639751
#> 869       0  1.656675158  1.174964842  0.453241313
#> 870       0 -2.022081286 -1.905372715 -1.723788786
#> 871       0  0.699734039 -0.005874030 -0.035326898
#> 872       0 -0.284292292 -0.390671052  0.608830823
#> 873       0 -1.167368621 -1.865114023 -2.006507850
#> 874       0  0.200417113  0.344140793  0.113582414
#> 875       0  0.316291286 -0.686068562  1.462430907
#> 876       0  0.623657245 -0.972542179  1.586366410
#> 877       0  0.253398410 -0.624238187 -0.142862516
#> 878       0 -0.346160022 -0.501127666  1.685246091
#> 879       0 -1.177133779 -2.036196837 -0.172807492
#> 880       0 -2.179599761 -2.712914254 -0.935532589
#> 881       0 -1.725707315 -2.820512900 -3.559906963
#> 882       0 -1.557369432 -2.772926068 -2.731656754
#> 883       0 -0.431383366 -1.168627410  0.094147384
#> 884       0  0.896131428  0.560881792  2.973083804
#> 885       0  0.097557484 -0.393229676  0.856685667
#> 886       0 -1.743314709 -3.082900705 -2.279216961
#> 887       0 -2.344488163 -1.435563283 -2.255204660
#> 888       0 -1.789435649 -1.498305031 -1.624173997
#> 889       0  0.140822667 -1.434405915  0.095045586
#> 890       0 -3.069242231 -3.712797683 -4.430320178
#> 891       0 -0.560277832 -1.907251402 -1.282782046
#> 892       0 -0.730197066 -1.924209626 -0.867176978
#> 893       0 -0.103882909 -0.546331814 -1.116750069
#> 894       0  0.498140800 -0.395424735 -0.443540996
#> 895       0  0.513260622  0.674104083  0.954496259
#> 896       0 -0.054786180 -1.124719294  0.533034913
#> 897       0  0.824376597  0.838675786  2.434167705
#> 898       0 -0.560277832 -0.064637904  1.738022171
#> 899       0 -0.440428486 -1.559090289 -2.522900649
#> 900       0  0.697135387 -0.838020739 -1.421399698
#> 901       0  0.793304107  2.024935105  0.876393816
#> 902       0 -2.004058295 -2.361166481 -3.336136856
#> 903       0 -1.005181934 -1.465181188 -1.620118314
#> 904       0 -1.756560548 -2.504233176 -1.298079833
#> 905       0 -1.314466013 -2.279801899 -0.194242511
#> 906       0 -0.440428486 -0.988067362 -0.539844056
#> 907       0  0.279666993  0.190043735  1.540468960
#> 908       0 -0.288401907 -2.072601121  0.377569531
#> 909       0 -1.317427229 -2.004000619 -0.518650912
#> 910       0  0.140822667 -1.168627410 -1.067862289
#> 911       0 -1.316093279 -1.754439190 -3.002032488
#> 912       0 -0.423314434 -1.375426561 -1.039456538
#> 913       0  0.489578511 -1.241887015  0.657435765
#> 914       0 -1.177133779 -1.498305031 -0.951787515
#> 915       0 -0.118696696 -1.566638938 -1.001299575
#> 916       0 -0.362287300 -1.498664881 -1.546853316
#> 917       0 -0.250388677  0.190043735  0.952588498
#> 918       0 -1.331568894 -1.955211876 -1.853927038
#> 919       0 -0.717063092 -1.275608741 -1.474532010
#> 920       0 -1.770521236 -2.038334754 -0.917686286
#> 921       0 -0.820177528 -1.582146965 -2.008219189
#> 922       0 -1.854995822 -2.496803403 -1.438337544
#> 923       0 -0.274171366 -0.922508612 -0.692028282
#> 924       0 -0.638266370 -1.955211876 -0.837267282
#> 925       0 -1.736567678 -2.390473070 -1.981547684
#> 926       0  0.279666993 -0.064637904  1.517489869
#> 927       0 -0.560277832 -0.229145813  0.631959853
#> 928       0 -0.274171366 -0.005874030  0.945444557
#> 929       0 -0.507195135 -1.840148830 -1.432336509
#> 930       0 -0.560277832 -1.314783591  1.250912889
#> 931       0 -2.927492507 -2.892119311 -2.289800236
#> 932       0  0.427196470  0.433596689  0.899829891
#> 933       0 -1.610681784 -1.443623594 -1.207835484
#> 934       0 -0.884406639 -0.440364450  0.465921575
#> 935       0 -0.637954351 -1.124719294 -0.280759713
#> 936       0  0.868146498  1.571927196  1.664871387
#> 937       0 -1.524549736 -0.370525777 -0.948986281
#> 938       0  0.896131428 -0.501127666  1.682221972
#> 939       0 -2.455354703 -3.118011273 -2.162776061
#> 940       0 -0.288401907 -0.624238187  0.587418523
#> 941       0 -1.331568894 -0.877088233  0.356856675
#> 942       0  0.065275309 -0.658202059 -0.317153475
#> 943       0 -1.445846363 -2.036196837 -1.481437987
#> 944       0 -0.560277832 -0.664386950 -1.990669444
#> 945       0  0.957810325  0.915731389  2.498366165
#> 946       0  0.280579357 -0.264326178  0.324320914
#> 947       0  1.806698793  1.084950383  0.895222646
#> 948       0  1.353816321  1.173632593  1.802367455
#> 949       0  0.595814141  1.926888054  0.364802173
#> 950       0  0.014562394 -0.342731345  0.381268872
#> 951       0 -0.672490266 -1.095405615  0.569697590
#> 952       0 -0.100453020 -0.818827187 -0.549087261
#> 953       0 -0.311265518 -0.682830213  0.166889733
#> 954       0 -1.317427229 -1.373614666 -0.911256296
#> 955       0  0.319238271 -0.559947522  0.213121286
#> 956       0  0.665089087  0.568883700  0.150221197
#> 957       0 -1.598016256 -2.266070448 -2.750619416
#> 958       0 -1.390880951 -1.755829534 -1.691135270
#> 959       0  0.243085799 -0.227891531 -1.326310399
#> 960       0 -0.600479913  0.288362252 -0.759343490
#> 961       0 -0.380612020 -1.420420046 -1.300735252
#> 962       0  0.474592506  0.036977119  0.829957649
#> 963       0  1.012320032  0.544826681  0.149567854
#> 964       0 -1.431190171 -1.498305031 -0.470509409
#> 965       0 -0.598164281 -0.962079122 -1.205855074
#> 966       0 -1.331568894 -1.942847492  0.617005511
#> 967       0 -0.834599062 -1.340114903 -1.219394858
#> 968       0 -1.317427229 -0.895118321  0.091312402
#> 969       0 -0.460064592  0.107841476  0.118477910
#> 970       0 -0.463552976 -0.549713195 -0.321729818
#> 971       0 -0.054786180 -1.998549577  1.218577031
#> 972       0 -1.006788858 -1.272298519  0.333778613
#> 973       0 -0.207325249 -0.790899405 -0.164120438
#> 974       0 -0.198671678 -1.424801374 -0.515737721
#> 975       0 -1.317427229 -1.215501088 -0.585724370
#> 976       0 -1.391352979 -2.276587625 -0.554345955
#> 977       0 -1.090163575 -2.504233176 -1.142323022
#> 978       0 -0.302652231 -0.403718017 -0.025212237
#> 979       0 -1.391352979 -1.942847492 -1.182021588
#> 980       0 -1.001776420 -1.999773280 -0.504013446
#> 981       0  0.953697789  0.506355585  0.524688786
#> 982       0 -0.930674470 -2.383798643 -3.030736973
#> 983       0  1.149041867 -0.124322538 -0.214029243
#> 984       0 -0.637954351 -0.556896818  0.531710091
#> 985       0 -2.004058295 -2.361166481 -2.195556550
#> 986       0 -0.869307914 -1.434405915 -0.231555284
#> 987       0  0.113279446  0.125406931  2.017461669
#> 988       0 -0.562262231 -1.331962265  1.075378490
#> 989       0 -0.638266370 -0.781372572  1.104165903
#> 990       0 -0.274171366 -0.730707398  1.739168447
#> 991       0  0.460179935 -0.490180198  0.100981820
#> 992       0  0.213541419 -1.815610955 -2.239312858
#> 993       0 -1.574354070 -1.072093592 -1.288697389
#> 994       0 -0.560277832 -1.168627410  2.041889298
#> 995       0 -1.238780503 -2.409644152 -3.071024177
#> 996       0  0.111594251  0.682887893 -2.155281612
#> 997       0 -0.048290580 -0.501127666 -0.789234909
#> 998       0 -1.304409643 -0.946489317  0.049963241
#> 999       0 -1.445846363 -1.942847492 -0.360933334
#> 1000      0 -0.867018060 -1.689133889 -0.004731789
#> 
#> [[3]]
#>      gender          edu         test        score
#> 801       0 -1.762386175 -1.211890953 -2.248523253
#> 802       0 -1.445846363 -2.072601121 -2.009755697
#> 803       0  1.170794530  1.169657867  1.191612986
#> 804       0  0.440124823 -0.244458389 -0.697736661
#> 805       0 -1.001776420 -0.390671052  1.984450007
#> 806       0 -0.430447496 -1.209261161  0.789841297
#> 807       0 -0.332573698 -1.180720819 -1.474926656
#> 808       0 -0.544864823  0.812795420  0.497834190
#> 809       0 -0.821708580 -1.353303876  0.982337819
#> 810       0 -1.996778408 -2.007616106 -1.670251654
#> 811       0 -0.346160022 -0.686068562  1.872163467
#> 812       0  0.113279446 -0.046012506  1.109306522
#> 813       0 -1.059961139 -0.364127628 -0.347495108
#> 814       0 -1.391352979 -1.215501088  0.841504242
#> 815       0 -0.424559369 -1.181002061 -1.031754054
#> 816       0 -1.090163575 -2.361166481 -1.038394788
#> 817       0 -1.317427229 -2.504233176 -1.253841133
#> 818       0 -1.317427229 -0.686068562 -1.127569653
#> 819       0 -1.177133779 -2.065701959 -1.321760978
#> 820       0  0.008667341 -0.985364068  1.062412448
#> 821       0 -1.770521236 -2.276587625 -1.627486071
#> 822       0 -0.592176006  0.124843632  0.871076451
#> 823       0  0.008667341 -0.212029256  0.667045683
#> 824       0 -0.289977458 -1.159948595 -1.783792760
#> 825       0  0.620924161 -0.114971666  0.050424648
#> 826       0 -1.307013160 -1.434405915 -1.244894150
#> 827       0 -0.313933684 -0.755109628  0.836772353
#> 828       0  0.566041719  0.807542567 -0.318456042
#> 829       0  1.663760353  1.480402631  2.299315681
#> 830       0 -1.789435649 -2.690683473 -0.780448031
#> 831       0 -1.090163575 -2.072601121  0.143680389
#> 832       0 -0.560277832 -0.839388153 -0.197988253
#> 833       0  0.675487598  0.868880270 -0.151364462
#> 834       0 -0.869229220 -2.646687427 -1.604124109
#> 835       0 -0.256472779 -1.440056662 -1.924364325
#> 836       0 -0.015333931 -0.820665352  1.861951160
#> 837       0 -1.770521236 -1.907251402 -1.462275080
#> 838       0  0.138359545 -1.790409294 -0.520312886
#> 839       0  0.229347734  0.106404515  0.318168927
#> 840       0  0.477895250  0.229617350 -0.422195985
#> 841       0 -1.268711009 -3.497560395 -0.094699039
#> 842       0  0.844084462 -0.305803234 -0.789176231
#> 843       0 -0.346160022 -0.756921041 -1.156295463
#> 844       0  0.799480614  0.389377080  0.988820130
#> 845       0 -0.094751785 -1.184286201 -0.383527997
#> 846       0  0.879040351 -0.046012506  1.297151367
#> 847       0 -1.043859083 -1.504998958 -0.509722245
#> 848       0  1.184813575  0.223671853  0.948738536
#> 849       0  0.509482542  0.031115903 -0.960078119
#> 850       0 -0.295124669 -0.004056565  0.069640746
#> 851       0 -0.280587143 -1.968535319 -0.642857333
#> 852       0  0.241140510  0.855783424  1.190722501
#> 853       0  0.799480614  0.198769381  1.672520558
#> 854       0 -1.043859083 -2.608594466 -1.376894259
#> 855       0  1.511649979  0.696805763  1.724080192
#> 856       0 -0.716919789  0.041877886  0.169921945
#> 857       0 -0.439514510 -2.782963224 -1.776423466
#> 858       0 -1.283178362 -1.390920653 -1.176246440
#> 859       0  0.279666993 -0.187664519  1.105173350
#> 860       0 -0.346160022 -1.912519623 -0.352586641
#> 861       0 -1.393405437 -2.227771994 -1.211904179
#> 862       0  0.650752035  0.870076964  0.172484884
#> 863       0 -0.257847380  0.184278366 -0.955779822
#> 864       0 -0.037313912 -0.556922603 -1.240644748
#> 865       0 -1.317427229 -0.559947522 -0.648786856
#> 866       0  0.836348371  1.801792786  0.816701249
#> 867       0 -0.659880020 -1.844452022 -0.468480289
#> 868       0  0.236144107  0.386751900  0.929639751
#> 869       0  1.656675158  1.174964842  0.453241313
#> 870       0 -2.022081286 -1.905372715 -1.723788786
#> 871       0  0.699734039 -0.005874030 -0.035326898
#> 872       0 -0.284292292 -0.390671052  0.608830823
#> 873       0 -1.167368621 -1.865114023 -2.006507850
#> 874       0  0.200417113  0.344140793  0.113582414
#> 875       0  0.316291286 -0.686068562  1.462430907
#> 876       0  0.623657245 -0.972542179  1.586366410
#> 877       0  0.253398410 -0.624238187 -0.142862516
#> 878       0 -0.346160022 -0.501127666  1.685246091
#> 879       0 -1.177133779 -2.036196837 -0.172807492
#> 880       0 -2.179599761 -2.712914254 -0.935532589
#> 881       0 -1.725707315 -2.820512900 -3.559906963
#> 882       0 -1.557369432 -2.772926068 -2.731656754
#> 883       0 -0.431383366 -1.168627410  0.094147384
#> 884       0  0.896131428  0.560881792  2.973083804
#> 885       0  0.097557484 -0.393229676  0.856685667
#> 886       0 -1.743314709 -3.082900705 -2.279216961
#> 887       0 -2.344488163 -1.435563283 -2.255204660
#> 888       0 -1.789435649 -1.498305031 -1.624173997
#> 889       0  0.140822667 -1.434405915  0.095045586
#> 890       0 -3.069242231 -3.712797683 -4.430320178
#> 891       0 -0.560277832 -1.907251402 -1.282782046
#> 892       0 -0.730197066 -1.924209626 -0.867176978
#> 893       0 -0.103882909 -0.546331814 -1.116750069
#> 894       0  0.498140800 -0.395424735 -0.443540996
#> 895       0  0.513260622  0.674104083  0.954496259
#> 896       0 -0.054786180 -1.124719294  0.533034913
#> 897       0  0.824376597  0.838675786  2.434167705
#> 898       0 -0.560277832 -0.064637904  1.738022171
#> 899       0 -0.440428486 -1.559090289 -2.522900649
#> 900       0  0.697135387 -0.838020739 -1.421399698
#> 901       0  0.793304107  2.024935105  0.876393816
#> 902       0 -2.004058295 -2.361166481 -3.336136856
#> 903       0 -1.005181934 -1.465181188 -1.620118314
#> 904       0 -1.756560548 -2.504233176 -1.298079833
#> 905       0 -1.314466013 -2.279801899 -0.194242511
#> 906       0 -0.440428486 -0.988067362 -0.539844056
#> 907       0  0.279666993  0.190043735  1.540468960
#> 908       0 -0.288401907 -2.072601121  0.377569531
#> 909       0 -1.317427229 -2.004000619 -0.518650912
#> 910       0  0.140822667 -1.168627410 -1.067862289
#> 911       0 -1.316093279 -1.754439190 -3.002032488
#> 912       0 -0.423314434 -1.375426561 -1.039456538
#> 913       0  0.489578511 -1.241887015  0.657435765
#> 914       0 -1.177133779 -1.498305031 -0.951787515
#> 915       0 -0.118696696 -1.566638938 -1.001299575
#> 916       0 -0.362287300 -1.498664881 -1.546853316
#> 917       0 -0.250388677  0.190043735  0.952588498
#> 918       0 -1.331568894 -1.955211876 -1.853927038
#> 919       0 -0.717063092 -1.275608741 -1.474532010
#> 920       0 -1.770521236 -2.038334754 -0.917686286
#> 921       0 -0.820177528 -1.582146965 -2.008219189
#> 922       0 -1.854995822 -2.496803403 -1.438337544
#> 923       0 -0.274171366 -0.922508612 -0.692028282
#> 924       0 -0.638266370 -1.955211876 -0.837267282
#> 925       0 -1.736567678 -2.390473070 -1.981547684
#> 926       0  0.279666993 -0.064637904  1.517489869
#> 927       0 -0.560277832 -0.229145813  0.631959853
#> 928       0 -0.274171366 -0.005874030  0.945444557
#> 929       0 -0.507195135 -1.840148830 -1.432336509
#> 930       0 -0.560277832 -1.314783591  1.250912889
#> 931       0 -2.927492507 -2.892119311 -2.289800236
#> 932       0  0.427196470  0.433596689  0.899829891
#> 933       0 -1.610681784 -1.443623594 -1.207835484
#> 934       0 -0.884406639 -0.440364450  0.465921575
#> 935       0 -0.637954351 -1.124719294 -0.280759713
#> 936       0  0.868146498  1.571927196  1.664871387
#> 937       0 -1.524549736 -0.370525777 -0.948986281
#> 938       0  0.896131428 -0.501127666  1.682221972
#> 939       0 -2.455354703 -3.118011273 -2.162776061
#> 940       0 -0.288401907 -0.624238187  0.587418523
#> 941       0 -1.331568894 -0.877088233  0.356856675
#> 942       0  0.065275309 -0.658202059 -0.317153475
#> 943       0 -1.445846363 -2.036196837 -1.481437987
#> 944       0 -0.560277832 -0.664386950 -1.990669444
#> 945       0  0.957810325  0.915731389  2.498366165
#> 946       0  0.280579357 -0.264326178  0.324320914
#> 947       0  1.806698793  1.084950383  0.895222646
#> 948       0  1.353816321  1.173632593  1.802367455
#> 949       0  0.595814141  1.926888054  0.364802173
#> 950       0  0.014562394 -0.342731345  0.381268872
#> 951       0 -0.672490266 -1.095405615  0.569697590
#> 952       0 -0.100453020 -0.818827187 -0.549087261
#> 953       0 -0.311265518 -0.682830213  0.166889733
#> 954       0 -1.317427229 -1.373614666 -0.911256296
#> 955       0  0.319238271 -0.559947522  0.213121286
#> 956       0  0.665089087  0.568883700  0.150221197
#> 957       0 -1.598016256 -2.266070448 -2.750619416
#> 958       0 -1.390880951 -1.755829534 -1.691135270
#> 959       0  0.243085799 -0.227891531 -1.326310399
#> 960       0 -0.600479913  0.288362252 -0.759343490
#> 961       0 -0.380612020 -1.420420046 -1.300735252
#> 962       0  0.474592506  0.036977119  0.829957649
#> 963       0  1.012320032  0.544826681  0.149567854
#> 964       0 -1.431190171 -1.498305031 -0.470509409
#> 965       0 -0.598164281 -0.962079122 -1.205855074
#> 966       0 -1.331568894 -1.942847492  0.617005511
#> 967       0 -0.834599062 -1.340114903 -1.219394858
#> 968       0 -1.317427229 -0.895118321  0.091312402
#> 969       0 -0.460064592  0.107841476  0.118477910
#> 970       0 -0.463552976 -0.549713195 -0.321729818
#> 971       0 -0.054786180 -1.998549577  1.218577031
#> 972       0 -1.006788858 -1.272298519  0.333778613
#> 973       0 -0.207325249 -0.790899405 -0.164120438
#> 974       0 -0.198671678 -1.424801374 -0.515737721
#> 975       0 -1.317427229 -1.215501088 -0.585724370
#> 976       0 -1.391352979 -2.276587625 -0.554345955
#> 977       0 -1.090163575 -2.504233176 -1.142323022
#> 978       0 -0.302652231 -0.403718017 -0.025212237
#> 979       0 -1.391352979 -1.942847492 -1.182021588
#> 980       0 -1.001776420 -1.999773280 -0.504013446
#> 981       0  0.953697789  0.506355585  0.524688786
#> 982       0 -0.930674470 -2.383798643 -3.030736973
#> 983       0  1.149041867 -0.124322538 -0.214029243
#> 984       0 -0.637954351 -0.556896818  0.531710091
#> 985       0 -2.004058295 -2.361166481 -2.195556550
#> 986       0 -0.869307914 -1.434405915 -0.231555284
#> 987       0  0.113279446  0.125406931  2.017461669
#> 988       0 -0.562262231 -1.331962265  1.075378490
#> 989       0 -0.638266370 -0.781372572  1.104165903
#> 990       0 -0.274171366 -0.730707398  1.739168447
#> 991       0  0.460179935 -0.490180198  0.100981820
#> 992       0  0.213541419 -1.815610955 -2.239312858
#> 993       0 -1.574354070 -1.072093592 -1.288697389
#> 994       0 -0.560277832 -1.168627410  2.041889298
#> 995       0 -1.238780503 -2.409644152 -3.071024177
#> 996       0  0.111594251  0.682887893 -2.155281612
#> 997       0 -0.048290580 -0.501127666 -0.789234909
#> 998       0 -1.304409643 -0.946489317  0.049963241
#> 999       0 -1.445846363 -1.942847492 -0.360933334
#> 1000      0 -0.867018060 -1.689133889 -0.004731789
#> 
#> [[4]]
#>      gender          edu         test        score
#> 801       0 -1.762386175 -1.211890953 -2.248523253
#> 802       0 -1.445846363 -2.072601121 -2.009755697
#> 803       0  1.170794530  1.169657867  1.191612986
#> 804       0  0.440124823 -0.244458389 -0.697736661
#> 805       0 -1.001776420 -0.390671052  1.984450007
#> 806       0 -0.430447496 -1.209261161  0.789841297
#> 807       0 -0.332573698 -1.180720819 -1.474926656
#> 808       0 -0.544864823  0.812795420  0.497834190
#> 809       0 -0.821708580 -1.353303876  0.982337819
#> 810       0 -1.996778408 -2.007616106 -1.670251654
#> 811       0 -0.346160022 -0.686068562  1.872163467
#> 812       0  0.113279446 -0.046012506  1.109306522
#> 813       0 -1.059961139 -0.364127628 -0.347495108
#> 814       0 -1.391352979 -1.215501088  0.841504242
#> 815       0 -0.424559369 -1.181002061 -1.031754054
#> 816       0 -1.090163575 -2.361166481 -1.038394788
#> 817       0 -1.317427229 -2.504233176 -1.253841133
#> 818       0 -1.317427229 -0.686068562 -1.127569653
#> 819       0 -1.177133779 -2.065701959 -1.321760978
#> 820       0  0.008667341 -0.985364068  1.062412448
#> 821       0 -1.770521236 -2.276587625 -1.627486071
#> 822       0 -0.592176006  0.124843632  0.871076451
#> 823       0  0.008667341 -0.212029256  0.667045683
#> 824       0 -0.289977458 -1.159948595 -1.783792760
#> 825       0  0.620924161 -0.114971666  0.050424648
#> 826       0 -1.307013160 -1.434405915 -1.244894150
#> 827       0 -0.313933684 -0.755109628  0.836772353
#> 828       0  0.566041719  0.807542567 -0.318456042
#> 829       0  1.663760353  1.480402631  2.299315681
#> 830       0 -1.789435649 -2.690683473 -0.780448031
#> 831       0 -1.090163575 -2.072601121  0.143680389
#> 832       0 -0.560277832 -0.839388153 -0.197988253
#> 833       0  0.675487598  0.868880270 -0.151364462
#> 834       0 -0.869229220 -2.646687427 -1.604124109
#> 835       0 -0.256472779 -1.440056662 -1.924364325
#> 836       0 -0.015333931 -0.820665352  1.861951160
#> 837       0 -1.770521236 -1.907251402 -1.462275080
#> 838       0  0.138359545 -1.790409294 -0.520312886
#> 839       0  0.229347734  0.106404515  0.318168927
#> 840       0  0.477895250  0.229617350 -0.422195985
#> 841       0 -1.268711009 -3.497560395 -0.094699039
#> 842       0  0.844084462 -0.305803234 -0.789176231
#> 843       0 -0.346160022 -0.756921041 -1.156295463
#> 844       0  0.799480614  0.389377080  0.988820130
#> 845       0 -0.094751785 -1.184286201 -0.383527997
#> 846       0  0.879040351 -0.046012506  1.297151367
#> 847       0 -1.043859083 -1.504998958 -0.509722245
#> 848       0  1.184813575  0.223671853  0.948738536
#> 849       0  0.509482542  0.031115903 -0.960078119
#> 850       0 -0.295124669 -0.004056565  0.069640746
#> 851       0 -0.280587143 -1.968535319 -0.642857333
#> 852       0  0.241140510  0.855783424  1.190722501
#> 853       0  0.799480614  0.198769381  1.672520558
#> 854       0 -1.043859083 -2.608594466 -1.376894259
#> 855       0  1.511649979  0.696805763  1.724080192
#> 856       0 -0.716919789  0.041877886  0.169921945
#> 857       0 -0.439514510 -2.782963224 -1.776423466
#> 858       0 -1.283178362 -1.390920653 -1.176246440
#> 859       0  0.279666993 -0.187664519  1.105173350
#> 860       0 -0.346160022 -1.912519623 -0.352586641
#> 861       0 -1.393405437 -2.227771994 -1.211904179
#> 862       0  0.650752035  0.870076964  0.172484884
#> 863       0 -0.257847380  0.184278366 -0.955779822
#> 864       0 -0.037313912 -0.556922603 -1.240644748
#> 865       0 -1.317427229 -0.559947522 -0.648786856
#> 866       0  0.836348371  1.801792786  0.816701249
#> 867       0 -0.659880020 -1.844452022 -0.468480289
#> 868       0  0.236144107  0.386751900  0.929639751
#> 869       0  1.656675158  1.174964842  0.453241313
#> 870       0 -2.022081286 -1.905372715 -1.723788786
#> 871       0  0.699734039 -0.005874030 -0.035326898
#> 872       0 -0.284292292 -0.390671052  0.608830823
#> 873       0 -1.167368621 -1.865114023 -2.006507850
#> 874       0  0.200417113  0.344140793  0.113582414
#> 875       0  0.316291286 -0.686068562  1.462430907
#> 876       0  0.623657245 -0.972542179  1.586366410
#> 877       0  0.253398410 -0.624238187 -0.142862516
#> 878       0 -0.346160022 -0.501127666  1.685246091
#> 879       0 -1.177133779 -2.036196837 -0.172807492
#> 880       0 -2.179599761 -2.712914254 -0.935532589
#> 881       0 -1.725707315 -2.820512900 -3.559906963
#> 882       0 -1.557369432 -2.772926068 -2.731656754
#> 883       0 -0.431383366 -1.168627410  0.094147384
#> 884       0  0.896131428  0.560881792  2.973083804
#> 885       0  0.097557484 -0.393229676  0.856685667
#> 886       0 -1.743314709 -3.082900705 -2.279216961
#> 887       0 -2.344488163 -1.435563283 -2.255204660
#> 888       0 -1.789435649 -1.498305031 -1.624173997
#> 889       0  0.140822667 -1.434405915  0.095045586
#> 890       0 -3.069242231 -3.712797683 -4.430320178
#> 891       0 -0.560277832 -1.907251402 -1.282782046
#> 892       0 -0.730197066 -1.924209626 -0.867176978
#> 893       0 -0.103882909 -0.546331814 -1.116750069
#> 894       0  0.498140800 -0.395424735 -0.443540996
#> 895       0  0.513260622  0.674104083  0.954496259
#> 896       0 -0.054786180 -1.124719294  0.533034913
#> 897       0  0.824376597  0.838675786  2.434167705
#> 898       0 -0.560277832 -0.064637904  1.738022171
#> 899       0 -0.440428486 -1.559090289 -2.522900649
#> 900       0  0.697135387 -0.838020739 -1.421399698
#> 901       0  0.793304107  2.024935105  0.876393816
#> 902       0 -2.004058295 -2.361166481 -3.336136856
#> 903       0 -1.005181934 -1.465181188 -1.620118314
#> 904       0 -1.756560548 -2.504233176 -1.298079833
#> 905       0 -1.314466013 -2.279801899 -0.194242511
#> 906       0 -0.440428486 -0.988067362 -0.539844056
#> 907       0  0.279666993  0.190043735  1.540468960
#> 908       0 -0.288401907 -2.072601121  0.377569531
#> 909       0 -1.317427229 -2.004000619 -0.518650912
#> 910       0  0.140822667 -1.168627410 -1.067862289
#> 911       0 -1.316093279 -1.754439190 -3.002032488
#> 912       0 -0.423314434 -1.375426561 -1.039456538
#> 913       0  0.489578511 -1.241887015  0.657435765
#> 914       0 -1.177133779 -1.498305031 -0.951787515
#> 915       0 -0.118696696 -1.566638938 -1.001299575
#> 916       0 -0.362287300 -1.498664881 -1.546853316
#> 917       0 -0.250388677  0.190043735  0.952588498
#> 918       0 -1.331568894 -1.955211876 -1.853927038
#> 919       0 -0.717063092 -1.275608741 -1.474532010
#> 920       0 -1.770521236 -2.038334754 -0.917686286
#> 921       0 -0.820177528 -1.582146965 -2.008219189
#> 922       0 -1.854995822 -2.496803403 -1.438337544
#> 923       0 -0.274171366 -0.922508612 -0.692028282
#> 924       0 -0.638266370 -1.955211876 -0.837267282
#> 925       0 -1.736567678 -2.390473070 -1.981547684
#> 926       0  0.279666993 -0.064637904  1.517489869
#> 927       0 -0.560277832 -0.229145813  0.631959853
#> 928       0 -0.274171366 -0.005874030  0.945444557
#> 929       0 -0.507195135 -1.840148830 -1.432336509
#> 930       0 -0.560277832 -1.314783591  1.250912889
#> 931       0 -2.927492507 -2.892119311 -2.289800236
#> 932       0  0.427196470  0.433596689  0.899829891
#> 933       0 -1.610681784 -1.443623594 -1.207835484
#> 934       0 -0.884406639 -0.440364450  0.465921575
#> 935       0 -0.637954351 -1.124719294 -0.280759713
#> 936       0  0.868146498  1.571927196  1.664871387
#> 937       0 -1.524549736 -0.370525777 -0.948986281
#> 938       0  0.896131428 -0.501127666  1.682221972
#> 939       0 -2.455354703 -3.118011273 -2.162776061
#> 940       0 -0.288401907 -0.624238187  0.587418523
#> 941       0 -1.331568894 -0.877088233  0.356856675
#> 942       0  0.065275309 -0.658202059 -0.317153475
#> 943       0 -1.445846363 -2.036196837 -1.481437987
#> 944       0 -0.560277832 -0.664386950 -1.990669444
#> 945       0  0.957810325  0.915731389  2.498366165
#> 946       0  0.280579357 -0.264326178  0.324320914
#> 947       0  1.806698793  1.084950383  0.895222646
#> 948       0  1.353816321  1.173632593  1.802367455
#> 949       0  0.595814141  1.926888054  0.364802173
#> 950       0  0.014562394 -0.342731345  0.381268872
#> 951       0 -0.672490266 -1.095405615  0.569697590
#> 952       0 -0.100453020 -0.818827187 -0.549087261
#> 953       0 -0.311265518 -0.682830213  0.166889733
#> 954       0 -1.317427229 -1.373614666 -0.911256296
#> 955       0  0.319238271 -0.559947522  0.213121286
#> 956       0  0.665089087  0.568883700  0.150221197
#> 957       0 -1.598016256 -2.266070448 -2.750619416
#> 958       0 -1.390880951 -1.755829534 -1.691135270
#> 959       0  0.243085799 -0.227891531 -1.326310399
#> 960       0 -0.600479913  0.288362252 -0.759343490
#> 961       0 -0.380612020 -1.420420046 -1.300735252
#> 962       0  0.474592506  0.036977119  0.829957649
#> 963       0  1.012320032  0.544826681  0.149567854
#> 964       0 -1.431190171 -1.498305031 -0.470509409
#> 965       0 -0.598164281 -0.962079122 -1.205855074
#> 966       0 -1.331568894 -1.942847492  0.617005511
#> 967       0 -0.834599062 -1.340114903 -1.219394858
#> 968       0 -1.317427229 -0.895118321  0.091312402
#> 969       0 -0.460064592  0.107841476  0.118477910
#> 970       0 -0.463552976 -0.549713195 -0.321729818
#> 971       0 -0.054786180 -1.998549577  1.218577031
#> 972       0 -1.006788858 -1.272298519  0.333778613
#> 973       0 -0.207325249 -0.790899405 -0.164120438
#> 974       0 -0.198671678 -1.424801374 -0.515737721
#> 975       0 -1.317427229 -1.215501088 -0.585724370
#> 976       0 -1.391352979 -2.276587625 -0.554345955
#> 977       0 -1.090163575 -2.504233176 -1.142323022
#> 978       0 -0.302652231 -0.403718017 -0.025212237
#> 979       0 -1.391352979 -1.942847492 -1.182021588
#> 980       0 -1.001776420 -1.999773280 -0.504013446
#> 981       0  0.953697789  0.506355585  0.524688786
#> 982       0 -0.930674470 -2.383798643 -3.030736973
#> 983       0  1.149041867 -0.124322538 -0.214029243
#> 984       0 -0.637954351 -0.556896818  0.531710091
#> 985       0 -2.004058295 -2.361166481 -2.195556550
#> 986       0 -0.869307914 -1.434405915 -0.231555284
#> 987       0  0.113279446  0.125406931  2.017461669
#> 988       0 -0.562262231 -1.331962265  1.075378490
#> 989       0 -0.638266370 -0.781372572  1.104165903
#> 990       0 -0.274171366 -0.730707398  1.739168447
#> 991       0  0.460179935 -0.490180198  0.100981820
#> 992       0  0.213541419 -1.815610955 -2.239312858
#> 993       0 -1.574354070 -1.072093592 -1.288697389
#> 994       0 -0.560277832 -1.168627410  2.041889298
#> 995       0 -1.238780503 -2.409644152 -3.071024177
#> 996       0  0.111594251  0.682887893 -2.155281612
#> 997       0 -0.048290580 -0.501127666 -0.789234909
#> 998       0 -1.304409643 -0.946489317  0.049963241
#> 999       0 -1.445846363 -1.942847492 -0.360933334
#> 1000      0 -0.867018060 -1.689133889 -0.004731789
#> 
#> [[5]]
#>      gender          edu         test        score
#> 801       0 -1.762386175 -1.211890953 -2.248523253
#> 802       0 -1.445846363 -2.072601121 -2.009755697
#> 803       0  1.170794530  1.169657867  1.191612986
#> 804       0  0.440124823 -0.244458389 -0.697736661
#> 805       0 -1.001776420 -0.390671052  1.984450007
#> 806       0 -0.430447496 -1.209261161  0.789841297
#> 807       0 -0.332573698 -1.180720819 -1.474926656
#> 808       0 -0.544864823  0.812795420  0.497834190
#> 809       0 -0.821708580 -1.353303876  0.982337819
#> 810       0 -1.996778408 -2.007616106 -1.670251654
#> 811       0 -0.346160022 -0.686068562  1.872163467
#> 812       0  0.113279446 -0.046012506  1.109306522
#> 813       0 -1.059961139 -0.364127628 -0.347495108
#> 814       0 -1.391352979 -1.215501088  0.841504242
#> 815       0 -0.424559369 -1.181002061 -1.031754054
#> 816       0 -1.090163575 -2.361166481 -1.038394788
#> 817       0 -1.317427229 -2.504233176 -1.253841133
#> 818       0 -1.317427229 -0.686068562 -1.127569653
#> 819       0 -1.177133779 -2.065701959 -1.321760978
#> 820       0  0.008667341 -0.985364068  1.062412448
#> 821       0 -1.770521236 -2.276587625 -1.627486071
#> 822       0 -0.592176006  0.124843632  0.871076451
#> 823       0  0.008667341 -0.212029256  0.667045683
#> 824       0 -0.289977458 -1.159948595 -1.783792760
#> 825       0  0.620924161 -0.114971666  0.050424648
#> 826       0 -1.307013160 -1.434405915 -1.244894150
#> 827       0 -0.313933684 -0.755109628  0.836772353
#> 828       0  0.566041719  0.807542567 -0.318456042
#> 829       0  1.663760353  1.480402631  2.299315681
#> 830       0 -1.789435649 -2.690683473 -0.780448031
#> 831       0 -1.090163575 -2.072601121  0.143680389
#> 832       0 -0.560277832 -0.839388153 -0.197988253
#> 833       0  0.675487598  0.868880270 -0.151364462
#> 834       0 -0.869229220 -2.646687427 -1.604124109
#> 835       0 -0.256472779 -1.440056662 -1.924364325
#> 836       0 -0.015333931 -0.820665352  1.861951160
#> 837       0 -1.770521236 -1.907251402 -1.462275080
#> 838       0  0.138359545 -1.790409294 -0.520312886
#> 839       0  0.229347734  0.106404515  0.318168927
#> 840       0  0.477895250  0.229617350 -0.422195985
#> 841       0 -1.268711009 -3.497560395 -0.094699039
#> 842       0  0.844084462 -0.305803234 -0.789176231
#> 843       0 -0.346160022 -0.756921041 -1.156295463
#> 844       0  0.799480614  0.389377080  0.988820130
#> 845       0 -0.094751785 -1.184286201 -0.383527997
#> 846       0  0.879040351 -0.046012506  1.297151367
#> 847       0 -1.043859083 -1.504998958 -0.509722245
#> 848       0  1.184813575  0.223671853  0.948738536
#> 849       0  0.509482542  0.031115903 -0.960078119
#> 850       0 -0.295124669 -0.004056565  0.069640746
#> 851       0 -0.280587143 -1.968535319 -0.642857333
#> 852       0  0.241140510  0.855783424  1.190722501
#> 853       0  0.799480614  0.198769381  1.672520558
#> 854       0 -1.043859083 -2.608594466 -1.376894259
#> 855       0  1.511649979  0.696805763  1.724080192
#> 856       0 -0.716919789  0.041877886  0.169921945
#> 857       0 -0.439514510 -2.782963224 -1.776423466
#> 858       0 -1.283178362 -1.390920653 -1.176246440
#> 859       0  0.279666993 -0.187664519  1.105173350
#> 860       0 -0.346160022 -1.912519623 -0.352586641
#> 861       0 -1.393405437 -2.227771994 -1.211904179
#> 862       0  0.650752035  0.870076964  0.172484884
#> 863       0 -0.257847380  0.184278366 -0.955779822
#> 864       0 -0.037313912 -0.556922603 -1.240644748
#> 865       0 -1.317427229 -0.559947522 -0.648786856
#> 866       0  0.836348371  1.801792786  0.816701249
#> 867       0 -0.659880020 -1.844452022 -0.468480289
#> 868       0  0.236144107  0.386751900  0.929639751
#> 869       0  1.656675158  1.174964842  0.453241313
#> 870       0 -2.022081286 -1.905372715 -1.723788786
#> 871       0  0.699734039 -0.005874030 -0.035326898
#> 872       0 -0.284292292 -0.390671052  0.608830823
#> 873       0 -1.167368621 -1.865114023 -2.006507850
#> 874       0  0.200417113  0.344140793  0.113582414
#> 875       0  0.316291286 -0.686068562  1.462430907
#> 876       0  0.623657245 -0.972542179  1.586366410
#> 877       0  0.253398410 -0.624238187 -0.142862516
#> 878       0 -0.346160022 -0.501127666  1.685246091
#> 879       0 -1.177133779 -2.036196837 -0.172807492
#> 880       0 -2.179599761 -2.712914254 -0.935532589
#> 881       0 -1.725707315 -2.820512900 -3.559906963
#> 882       0 -1.557369432 -2.772926068 -2.731656754
#> 883       0 -0.431383366 -1.168627410  0.094147384
#> 884       0  0.896131428  0.560881792  2.973083804
#> 885       0  0.097557484 -0.393229676  0.856685667
#> 886       0 -1.743314709 -3.082900705 -2.279216961
#> 887       0 -2.344488163 -1.435563283 -2.255204660
#> 888       0 -1.789435649 -1.498305031 -1.624173997
#> 889       0  0.140822667 -1.434405915  0.095045586
#> 890       0 -3.069242231 -3.712797683 -4.430320178
#> 891       0 -0.560277832 -1.907251402 -1.282782046
#> 892       0 -0.730197066 -1.924209626 -0.867176978
#> 893       0 -0.103882909 -0.546331814 -1.116750069
#> 894       0  0.498140800 -0.395424735 -0.443540996
#> 895       0  0.513260622  0.674104083  0.954496259
#> 896       0 -0.054786180 -1.124719294  0.533034913
#> 897       0  0.824376597  0.838675786  2.434167705
#> 898       0 -0.560277832 -0.064637904  1.738022171
#> 899       0 -0.440428486 -1.559090289 -2.522900649
#> 900       0  0.697135387 -0.838020739 -1.421399698
#> 901       0  0.793304107  2.024935105  0.876393816
#> 902       0 -2.004058295 -2.361166481 -3.336136856
#> 903       0 -1.005181934 -1.465181188 -1.620118314
#> 904       0 -1.756560548 -2.504233176 -1.298079833
#> 905       0 -1.314466013 -2.279801899 -0.194242511
#> 906       0 -0.440428486 -0.988067362 -0.539844056
#> 907       0  0.279666993  0.190043735  1.540468960
#> 908       0 -0.288401907 -2.072601121  0.377569531
#> 909       0 -1.317427229 -2.004000619 -0.518650912
#> 910       0  0.140822667 -1.168627410 -1.067862289
#> 911       0 -1.316093279 -1.754439190 -3.002032488
#> 912       0 -0.423314434 -1.375426561 -1.039456538
#> 913       0  0.489578511 -1.241887015  0.657435765
#> 914       0 -1.177133779 -1.498305031 -0.951787515
#> 915       0 -0.118696696 -1.566638938 -1.001299575
#> 916       0 -0.362287300 -1.498664881 -1.546853316
#> 917       0 -0.250388677  0.190043735  0.952588498
#> 918       0 -1.331568894 -1.955211876 -1.853927038
#> 919       0 -0.717063092 -1.275608741 -1.474532010
#> 920       0 -1.770521236 -2.038334754 -0.917686286
#> 921       0 -0.820177528 -1.582146965 -2.008219189
#> 922       0 -1.854995822 -2.496803403 -1.438337544
#> 923       0 -0.274171366 -0.922508612 -0.692028282
#> 924       0 -0.638266370 -1.955211876 -0.837267282
#> 925       0 -1.736567678 -2.390473070 -1.981547684
#> 926       0  0.279666993 -0.064637904  1.517489869
#> 927       0 -0.560277832 -0.229145813  0.631959853
#> 928       0 -0.274171366 -0.005874030  0.945444557
#> 929       0 -0.507195135 -1.840148830 -1.432336509
#> 930       0 -0.560277832 -1.314783591  1.250912889
#> 931       0 -2.927492507 -2.892119311 -2.289800236
#> 932       0  0.427196470  0.433596689  0.899829891
#> 933       0 -1.610681784 -1.443623594 -1.207835484
#> 934       0 -0.884406639 -0.440364450  0.465921575
#> 935       0 -0.637954351 -1.124719294 -0.280759713
#> 936       0  0.868146498  1.571927196  1.664871387
#> 937       0 -1.524549736 -0.370525777 -0.948986281
#> 938       0  0.896131428 -0.501127666  1.682221972
#> 939       0 -2.455354703 -3.118011273 -2.162776061
#> 940       0 -0.288401907 -0.624238187  0.587418523
#> 941       0 -1.331568894 -0.877088233  0.356856675
#> 942       0  0.065275309 -0.658202059 -0.317153475
#> 943       0 -1.445846363 -2.036196837 -1.481437987
#> 944       0 -0.560277832 -0.664386950 -1.990669444
#> 945       0  0.957810325  0.915731389  2.498366165
#> 946       0  0.280579357 -0.264326178  0.324320914
#> 947       0  1.806698793  1.084950383  0.895222646
#> 948       0  1.353816321  1.173632593  1.802367455
#> 949       0  0.595814141  1.926888054  0.364802173
#> 950       0  0.014562394 -0.342731345  0.381268872
#> 951       0 -0.672490266 -1.095405615  0.569697590
#> 952       0 -0.100453020 -0.818827187 -0.549087261
#> 953       0 -0.311265518 -0.682830213  0.166889733
#> 954       0 -1.317427229 -1.373614666 -0.911256296
#> 955       0  0.319238271 -0.559947522  0.213121286
#> 956       0  0.665089087  0.568883700  0.150221197
#> 957       0 -1.598016256 -2.266070448 -2.750619416
#> 958       0 -1.390880951 -1.755829534 -1.691135270
#> 959       0  0.243085799 -0.227891531 -1.326310399
#> 960       0 -0.600479913  0.288362252 -0.759343490
#> 961       0 -0.380612020 -1.420420046 -1.300735252
#> 962       0  0.474592506  0.036977119  0.829957649
#> 963       0  1.012320032  0.544826681  0.149567854
#> 964       0 -1.431190171 -1.498305031 -0.470509409
#> 965       0 -0.598164281 -0.962079122 -1.205855074
#> 966       0 -1.331568894 -1.942847492  0.617005511
#> 967       0 -0.834599062 -1.340114903 -1.219394858
#> 968       0 -1.317427229 -0.895118321  0.091312402
#> 969       0 -0.460064592  0.107841476  0.118477910
#> 970       0 -0.463552976 -0.549713195 -0.321729818
#> 971       0 -0.054786180 -1.998549577  1.218577031
#> 972       0 -1.006788858 -1.272298519  0.333778613
#> 973       0 -0.207325249 -0.790899405 -0.164120438
#> 974       0 -0.198671678 -1.424801374 -0.515737721
#> 975       0 -1.317427229 -1.215501088 -0.585724370
#> 976       0 -1.391352979 -2.276587625 -0.554345955
#> 977       0 -1.090163575 -2.504233176 -1.142323022
#> 978       0 -0.302652231 -0.403718017 -0.025212237
#> 979       0 -1.391352979 -1.942847492 -1.182021588
#> 980       0 -1.001776420 -1.999773280 -0.504013446
#> 981       0  0.953697789  0.506355585  0.524688786
#> 982       0 -0.930674470 -2.383798643 -3.030736973
#> 983       0  1.149041867 -0.124322538 -0.214029243
#> 984       0 -0.637954351 -0.556896818  0.531710091
#> 985       0 -2.004058295 -2.361166481 -2.195556550
#> 986       0 -0.869307914 -1.434405915 -0.231555284
#> 987       0  0.113279446  0.125406931  2.017461669
#> 988       0 -0.562262231 -1.331962265  1.075378490
#> 989       0 -0.638266370 -0.781372572  1.104165903
#> 990       0 -0.274171366 -0.730707398  1.739168447
#> 991       0  0.460179935 -0.490180198  0.100981820
#> 992       0  0.213541419 -1.815610955 -2.239312858
#> 993       0 -1.574354070 -1.072093592 -1.288697389
#> 994       0 -0.560277832 -1.168627410  2.041889298
#> 995       0 -1.238780503 -2.409644152 -3.071024177
#> 996       0  0.111594251  0.682887893 -2.155281612
#> 997       0 -0.048290580 -0.501127666 -0.789234909
#> 998       0 -1.304409643 -0.946489317  0.049963241
#> 999       0 -1.445846363 -1.942847492 -0.360933334
#> 1000      0 -0.867018060 -1.689133889 -0.004731789
#>